mental_health_diary/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework_simplejwt==5.3.1
django-cors-headers==4.6.0
django-extensions==3.2.3
pillow==11.1.0
whitenoise==6.9.0